tools/QueryToTax/requirements.txt
elasticsearch_dsl 
pandas
scikit-learn
joblib
